#226 - Code changes for consistency
