+++ src/controls/iFramePanel/IFramePanelContentProps.tsx
@@ -1,0 +1,4 @@
+export interface IIFramePanelContentProps extends React.IframeHTMLAttributes<HTMLIFrameElement> {
+  close: () => void;
+  iframeOnLoad?: (iframe: any) => void;
+}
