+++ src/controls/iFramePanel/IFramePanelContentState.tsx
@@ -1,0 +1,3 @@
+export interface IIFramePanelContentState {
+  isContentVisible?: boolean;
+}
