+++ src/controls/iFramePanel/IFramePanelState.ts
@@ -1,0 +1,2 @@
+export interface IIFramePanelState {
+}
