+++ src/controls/iFramePanel/index.ts
@@ -1,3 +1,7 @@
 export * from './IFramePanelContent.module.scss';
 export * from './IFramePanelContent';
+export * from './IFramePanelContentProps';
+export * from './IFramePanelContentState';
 export * from './IFramePanel';
+export * from './IFramePanelProps';
+export * from './IFramePanelState';
